--- app/controllers/transactions_controller.rb
@@ -1,88 +1,0 @@
-class TransactionsController < ApplicationController
-  before_action :set_transaction, only: %i[show edit update destroy]
-
-  def new
-    @transaction = Transaction.new
-  end
-
-  def edit
-    authorize @transaction
-  end
-
-  def show
-    authorize @transaction
-  end
-
-  def create
-    @transaction = Transactions::CreateTransaction.perform(transactions_params)
-    if @transaction.save
-      respond_to do |format|
-        format.html do
-          redirect_to transaction_path(@transaction),
-                      notice: 'Transaction successfully created.'
-        end
-        format.json { render json: @transaction, status: :created }
-      end
-
-    else
-      respond_to do |format|
-        format.html { render :new }
-        format.json { render json: @transaction.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  def index
-    @transactions = policy_scope(Transaction).current_month.order(date: :desc)
-  end
-
-  def update
-    authorize @transaction
-
-    if @transaction.update(transactions_params)
-      respond_to do |format|
-        format.html do
-          redirect_to @transaction,
-                      notice: 'Transaction successfully updated.'
-        end
-        format.json { render json: @transaction, status: :ok }
-      end
-
-    else
-      respond_to do |format|
-        format.html { render :edit }
-        format.json { render json: @transaction.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  def destroy
-    authorize @transaction
-
-    if @transaction.destroy
-      respond_to do |format|
-        format.html do
-          redirect_to transactions_path(id: params[:account_id]),
-                      notice: 'Transaction successfully removed.'
-        end
-        format.json { head :no_content, status: :ok }
-      end
-    else
-      respond_to do |format|
-        format.html { render :delete }
-        format.json { render json: @transaction.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  private
-
-  def transactions_params
-    params.require(:transaction).permit(:title, :account_id, :category_id,
-                                        :value, :kind, :date)
-  end
-
-  def set_transaction
-    @transaction = Transaction.find(params[:id])
-  end
-end
